NOJIRA Remove unused method for submitting messages in bulk to SMPP gateway.
Although submit_multi is supported by the SMPP spec, as at Sep 09 this is not supported by the Clickatell SMPP service, so this will not be implemented in the near future.

In any event the code support for this was not complete. The method would return a single message id for a set of messages, and delivery reports would then need to be matched against both message id and mobile number, rather than only message id.